Updates to new React Native MainApplication code
The getPackages method has changed with the latest version of React Native to not include the default firebase package and requires a different way of adding module packages. This updates the documentation for Auth to reflect this. If this PR is approved, I can make this same change across other modules.